Refactor packaging
This commit will adjust packaging and package names according to REST
architecture best practices.

Also, the application properties file extension was change for better
readability.
