+++ pom.xml
@@ -11,7 +11,7 @@
 	<groupId>br.com.codenation</groupId>
 	<artifactId>squad-1-ad-java-softplan-2</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>squad-1-ad-java-softplan-2</name>
+	<name>SoftLog</name>
 	<description>Gerenciador Log</description>
 
 	<properties>
@@ -19,6 +19,7 @@
 	</properties>
 
 	<dependencies>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
@@ -62,18 +63,20 @@
 		<dependency>
 			<groupId>org.flywaydb</groupId>
 			<artifactId>flyway-core</artifactId>
-    </dependency>
+    	</dependency>
     
 		<dependency>
 			<groupId>io.springfox</groupId>
 			<artifactId>springfox-swagger2</artifactId>
 			<version>2.9.2</version>
 		</dependency>
+
 		<dependency>
 			<groupId>io.springfox</groupId>
 			<artifactId>springfox-swagger-ui</artifactId>
 			<version>2.9.2</version>
 		</dependency>
+
 	</dependencies>
 
 	<build>
