--- src/main/java/br/com/codenation/squad1/auth/application/AuthService.java
@@ -1,5 +1,0 @@
-package br.com.codenation.squad1.auth.application;
-
-public class AuthService {
-
-}
