+++ src/main/java/br/com/codenation/squad1/auth/service/AuthService.java
@@ -1,0 +1,5 @@
+package br.com.codenation.squad1.auth.service;
+
+public class AuthService {
+
+}
