--- src/main/java/br/com/codenation/squad1/log/domain/enums/Status.java
@@ -1,7 +1,0 @@
-package br.com.codenation.squad1.log.domain.enums;
-
-public enum Status {
-
-	HIDDEN, DELETED
-
-}
