+++ src/main/java/br/com/codenation/squad1/log/model/enums/Status.java
@@ -1,0 +1,7 @@
+package br.com.codenation.squad1.log.model.enums;
+
+public enum Status {
+
+	HIDDEN, DELETED
+
+}
