+++ src/main/java/br/com/codenation/squad1/user/repository/UserAccountRepository.java
@@ -1,8 +1,8 @@
-package br.com.codenation.squad1.user.infra.repository;
+package br.com.codenation.squad1.user.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import br.com.codenation.squad1.user.domain.UserAccount;
+import br.com.codenation.squad1.user.model.UserAccount;
 
 public interface UserAccountRepository extends CrudRepository<UserAccount, Long> {
