--- src/main/java/br/com/codenation/squad1/util/PasswordManager.java
@@ -1,9 +1,0 @@
-package br.com.codenation.squad1.util;
-
-public class PasswordManager {
-
-    public static void main(String[] args) {
-//        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-//        System.out.println(encoder.encode("admin"));
-    }
-}
