--- src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/squad1_db
-spring.datasource.username= postgres
-spring.datasource.password= postgres
-
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.show-sql=true
